docs(reference): add docs section for File Resolution
Refs #361
